bot_maekup_llm-master/requirements.txt
pyTelegramBotAPI
huggingface_hub
psycopg2-binary
